+++ CHANGELOG.md
@@ -9,11 +9,12 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 - Updated elastic-lib v4.2.0
 
 ### Added
-- get_dumps: 
+- get_dumps: Retrieve dumps from the Elasticsearch cluster and return the dumps in a dictionary format, includes the option to only return failed dumps.
 - create_header: Create standard dictionary header.
 
 ### Changed
-- list_dumps: Converted output to a dictionary format and also call data_out function to handle the output.
+- failed_dumps: Added check for existing repo name, converted output to a dictionary format and call data_out function to handle the output.
+- list_dumps: Converted output to a dictionary format and call data_out function to handle the output.
 - data_out: Refactored the function to be more stream-lined.
 - run_program: Created gen_class.TimeFormat instance and passed to functions being called.
 - run_program: Removed "port" and "scheme" variables from the elastic_class.ElasticSearchStatus instance call.
@@ -24,6 +25,7 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 - Support for Elasticsearch v7.4, v7.12 and v7.17
 
 ### Removed
+- print_failures function
 - print_dumps function
 - Support for Elasticsearch v6.4